Refactors and reorganizes test utilities
Renames and consolidates test-related files for clarity and consistency.
Adds an index file to streamline imports for test utilities.
Introduces a debug log for improved test diagnostics.

Improves maintainability and developer experience.
