+++ src/lint/colors/tests/index.ts
@@ -1,0 +1,5 @@
+import { astroColorIsUsingCorrectTheme } from "./astro-color-is-using-correct-theme";
+import { usingAstroColor } from "./using-astro-color";
+import { usingColorFromComponent } from "./using-color-from-component";
+
+export { usingAstroColor, usingColorFromComponent, astroColorIsUsingCorrectTheme };
